:wrench: Renamed modules, packages and two main classes
