--- extended-hateotas-common/extended-hateotas-common-shared/src/main/java/com/github/forinil/extendedhateotas/model/Right.java
@@ -1,9 +1,0 @@
-package com.github.forinil.extendedhateotas.model;
-
-public enum Right {
-    R_SAVE,
-    R_SIGN,
-    R_SEND,
-    R_DELETE,
-    R_LIST
-}
