--- extended-hateotas-common/pom.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<artifactId>extended-hateotas-common</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<packaging>pom</packaging>
-
-	<name>extended-hateotas-common</name>
-	<description>Extended HATEOTAS Proof-of-concept common libraries</description>
-
-	<parent>
-		<groupId>com.github.forinil</groupId>
-		<artifactId>extended-hateotas</artifactId>
-		<version>0.0.1-SNAPSHOT</version>
-	</parent>
-
-	<modules>
-		<module>extended-hateotas-common-shared</module>
-		<module>extended-hateotas-common-data</module>
-	</modules>
-</project>
