--- extended-hateotas-facade/src/main/java/com/github/forinil/extendedhateotas/model/Action.java
@@ -1,9 +1,0 @@
-package com.github.forinil.extendedhateotas.model;
-
-public enum Action {
-    SAVE,
-    SIGN,
-    SEND,
-    DELETE,
-    LIST
-}
