+++ hateoas-dual-layer-common/hateoas-dual-layer-common-rest/src/main/java/com/github/forinil/hateoasduallayer/model/ApplicationData.java
@@ -1,0 +1,29 @@
+package com.github.forinil.hateoasduallayer.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.*;
+import org.springframework.hateoas.ResourceSupport;
+
+import java.util.Date;
+
+@NoArgsConstructor
+@ToString
+public class ApplicationData extends ResourceSupport {
+
+    @Getter
+    @Setter
+    @JsonProperty("id")
+    private int applicationId;
+
+    @NonNull
+    @Getter
+    @Setter
+    @JsonProperty
+    private Date creationDate;
+
+    @NonNull
+    @Getter
+    @Setter
+    @JsonProperty
+    private String content;
+}
