+++ hateoas-dual-layer-common/hateoas-dual-layer-common-rest/src/main/java/com/github/forinil/hateoasduallayer/model/Right.java
@@ -1,0 +1,9 @@
+package com.github.forinil.hateoasduallayer.model;
+
+public enum Right {
+    R_SAVE,
+    R_SIGN,
+    R_SEND,
+    R_DELETE,
+    R_LIST
+}
