+++ hateoas-dual-layer-frontend-facade/src/main/java/com/github/forinil/hateoasduallayer/model/Action.java
@@ -1,0 +1,9 @@
+package com.github.forinil.hateoasduallayer.model;
+
+public enum Action {
+    SAVE,
+    SIGN,
+    SEND,
+    DELETE,
+    LIST
+}
