+++ hateoas-dual-layer-frontend-facade/src/main/resources/application.yml
